simplify non-flake usage (#28)
* simplify non-flake usage

This change breaks backward-compatibility.

* package.nix becomes default.nix
* overlay.nix disappears. Build your own overlay if you don't use
  flakes.
* replace flake-compat with flake.lock.nix that just tries to pull the
  dependencies from the lockfile.

This brings back the convention that default.nix holds all of the
packages. The main advantage is that non-flake users can now also pass
their own versions of everything to the constructore. A secondary
advantage is that `nix repl .` works and makes it easier to debug. A
third smaller advantage is that the flake.nix is now properly lean.

* Update default.nix

Co-authored-by: figsoda <figsoda@pm.me>

* remove unused flake.lock.nix

* rename nixpkgs input to pkgs to make it callPackage friendly

* use genAttrs in flake.nix

* fix flake check

* Revert "fix flake check"

This reverts commit a48a2e7a0ed0883fd4bd0a232459d06440bc0403.

* fmt

* add overlay.nix

Co-authored-by: figsoda <figsoda@pm.me>