don't take a `system` argument in `default.nix`
Just one `system` is insufficient for cross compilation, although I'm
not sure cross compilation is a concern for fenix.

Either way, nixpkgs' default behavior is good enough, and if callers
need to do special things about systems, they can do so by providing an
instance of nixpkgs configured the way they need it to be.