Always patch the interpreter of rustlib binaries (#131)
Closes #206 
Closes #211 

When using `stable.toolchain`,
`lib/rustlib/x86_64-unknown-linux-gnu/bin/rust-lld` and other binaries
in the same directory are not patched, so they do not work. I'm not sure
exactly when Rust uses these, but it seems to break cross-compilation to
`wasm32-unknown-unknown`.

Fix it by generalizing the patching applied to the `rustc` component.
Replace the glob with a `find` invocation to avoid errors on components
that have a `lib` directory but no `lib/rustlib/*/bin`.